Vulkan: Add r_shaderGamma as alternative to HW gamma
SetDeviceGammaRamp Win32 API shows its age. I experienced issues in
multi-monitor configuration. HW gamma also affects screen globally -
that's a not great experience when you run the game in a window.

r_shaderGamma emulates exactly what HW gamma does in the compute
shader (no difference when switching between two modes).

This is supported only by the Vulkan backend. No plans to use this for
OpenGL, the goal of this project is to preserve the original behavior
of the game, only minor adjustment are made but not a big one as
switching to new OpenGL version with modern features. Vulkan backend
is a safety net, since OpenGL support is getting worse, I feel that
these days it provides better experience.