Vulkan: Add vsync support (r_vsync cvar)
NOTE: the recommendation is to set com_maxFPS to 0 if monitor refresh
rate is higher than the current com_maxFPS value. Otherwise, com_maxFPS
will prevent from keeping high enough framerate to keep up with vsync events.

For 60hz monitors it is recommended not to use vsync and just set
com_maxFPS to 125 (common value, but can be different). 60 FPS is not
enough for smooth Q3 gameplay even if monitor is only 60hz.